Added support for Blaze Serialization, cleanup

+++ src/Skate3Server.Blaze/BlazeHeader.cs
@@ -2,13 +2,13 @@
 {
     public class BlazeHeader
     {
-        public int Length { get; set; }
+        public ushort Length { get; set; }
         
         public BlazeComponent Component { get; set; }
         
-        public int Command { get; set; }
+        public ushort Command { get; set; }
         
-        public int ErrorCode { get; set; }
+        public ushort ErrorCode { get; set; }
 
         public BlazeMessageType MessageType { get; set; }
